[S09P22B310-000] feat : makeTransaction추가

+++ backend/src/main/java/com/ntt/mwonimoney/domain/account/api/response/FinAccountResponse.java
@@ -10,7 +10,7 @@ import lombok.Data;
 @Data
 public class FinAccountResponse {
 	private String number;
-	private Long remain;
+	private int remain;
 	private LocalDate createdDay;
 	private String status;
 	List<FinAccountTransactionDto> finAccountTransaction;
