+++ backend/src/main/java/com/ntt/mwonimoney/domain/account/model/dto/FinAccountDto.java
@@ -17,7 +17,7 @@ public class FinAccountDto {
 
 	private FinAccountType type;
 
-	private Long remain;
+	private int remain;
 
 }
